Make (a+b+c+...)^n work for huuuge terms.
This fix is neccessary for results exceeding 2^31 terms. While at it,
also restructured power::expand_add_() a bit to look more like the
more general power::expand_add().

+++ ginac/power.h
@@ -95,7 +95,7 @@ protected:
 	void do_print_python_repr(const print_python_repr & c, unsigned level) const;
 	void do_print_csrc_cl_N(const print_csrc_cl_N & c, unsigned level) const;
 
-	ex expand_add(const add & a, int n, unsigned options) const;
+	ex expand_add(const add & a, long n, unsigned options) const;
 	ex expand_add_2(const add & a, unsigned options) const;
 	ex expand_mul(const mul & m, const numeric & n, unsigned options, bool from_expand = false) const;
